Move players and its components to `Players/` dir
