+++ app/src/App.js
@@ -16,8 +16,7 @@ import AppMenu from './AppMenu';
 import ModalRouter from './ModalRouter';
 import GAListener from './GAListener';
 import ErrorCatcher from './ErrorCatcher';
-import Rac1Directe from './Rac1Directe';
-import Rac1ByDate from './Rac1ByDate';
+import { Rac1Directe, Rac1ByDate } from './Players';
 import Cookies from './Cookies';
 import About from './About';
 import Help from './Help';
